Add CentOS image support
Each image has it's own Dockerfile, so when building the image, don't forget to
specify it as follows:

$ docker build -f Dockerfile.debian .

Script `get-python.sh` is now extended to be distribution-agnostic via shell
`case` statements. You can add support for more distributions by complementing
the cases, but don't forget to test the result. For now Debian and CentOS are
supported.