Drop python 2.6
Well, it's been awhile since Python 2.6 reached its end of support. The
only reason why it was a part of Pythonista image is because I used it.
Fortunately, I don't need it anymore, so I can remove it from image.
I want this image to be focused on all *live* Python versions.