Check PyPy's version inside container
So far we checked only Python version PyPy compatible with. It's ok for
basic check, and is not good to make sure PyPy was really updated.

For example, both PyPy 4.0.1 and 5.0.1 are compatible with 2.7.10. Thus,
when we update PyPy we can't ensure it's really updated.

That commit introduces additional check for PyPy version.